Add README_CN, DNS benchmark and fixes
Add a Chinese README and new DNS benchmark screen, register new routes, and apply multiple bug fixes and refactors across the app.

Changes include:
- Added README_CN.md (Chinese project documentation).
- New DnsBenchmarkScreen composable and NavGraph entries; also added PasswordStrengthChecker route.
- CompressImage refactor: renamed enum to ImageOutputFormat, updated UI (FormatSelector), fixed batch save flow (pendingNextSave + LaunchedEffect), and updated compressBitmap signature/usage; switched several icons to AutoMirrored.
- ApiTester: read response body bytes non-null, compute body size accordingly; replaced ScrollableTabRow with PrimaryScrollableTabRow.
- CdnLatencyTest: removed nullable Socket usage and cleaned connect/close calls.
- LocaleHelper: added deprecation suppression and simplified createLocale logic.
- CustomRequest: minor nullability/header handling fix.
- build.gradle.kts: removed kotlin android plugin alias in app module.
- Updated .claude/settings.local.json permissions and many localized strings (values/*/strings.xml).

Mostly feature additions and stability/UX fixes to improve correctness and internationalization.

+++ app/src/main/java/cv/toolkit/screens/SubnetCalculatorScreen.kt
@@ -89,7 +89,7 @@ fun SubnetCalculatorScreen(navController: NavController) {
         }
 
         // Calculate subnet mask
-        val maskBits = (0xFFFFFFFF.toUInt() shl (32 - cidrValue)).toUInt()
+        val maskBits = 0xFFFFFFFF.toUInt() shl (32 - cidrValue)
         val subnetMask = listOf(
             ((maskBits shr 24) and 0xFFu).toInt(),
             ((maskBits shr 16) and 0xFFu).toInt(),
